pyproject.toml
[project]
name = "document-organizer"
version = "0.1.0"
description = "Automatically organize scanned documents using AI-powered categorization with the Johnny.Decimal system"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Document Organizer Contributors"}
]
keywords = ["document", "organizer", "ai", "johnny-decimal", "pdf", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business",
    "Topic :: Utilities",
]

dependencies = [
    "docling>=2.0.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "streamlit>=1.40.0",
    "requests>=2.31.0",  # For Ollama provider
]

[project.optional-dependencies]
# AI Provider extras - install what you need
anthropic = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
bedrock = ["boto3>=1.34.0"]
ollama = ["requests>=2.31.0"]

# Install all AI providers
all-providers = [
    "anthropic>=0.40.0",
    "openai>=1.0.0",
    "boto3>=1.34.0",
]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
]

[project.scripts]
document-organizer = "document_organizer:main"

[project.urls]
Homepage = "https://github.com/yourusername/document-organizer"
Repository = "https://github.com/yourusername/document-organizer"
Issues = "https://github.com/yourusername/document-organizer/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["document_organizer", "ai_providers", "ui", "migrate_to_jd", "preview_renames"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "venv/*", "*_env/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]
